Documented theme and feature flag services
